finished movement (for now)
- Tanks have functioning interactions with each other
- pushing another Tank will slow you down, while (possibly) speeding up
  the other one
- Walls, Breakables, and Holes each use Solid.cs to indicate movement
  interaction and cost

+++ Assets/scripts/Interfaces.cs
@@ -3,7 +3,7 @@ interface MovementInteraction
 {
     float moveX(float movx);
     float moveY(float movy);
-    int movementCost();
+    float movementCost();
 }
 
 interface BombInteraction
